Improved Movement
Player moves only in cardinal directions now.
Player moves in the direction of the latest keypress.
If multiple keys are pressed, the player moves in the direction of the latest keypress. If that key is then unpressed, the player will resume moving in the direction of the first key that was overwritten.