Euclidean Pathing Algorithm + AI Intelligence Levels
Euclidiean 'as the crow flies' algorithm to minimise distance to target.
Intelligence levels now affect AI pathing. A 'fail' will cause the AI to continue moving in the direction it was before, rather than in the desired direction, this means they will sometimes fail to follow you through an intersection correctly.