PlayerCharacter enum and Rectangles!
the PlayerCharacter enum holds all the characters the player can be.
Fixed a recurring bug where LevelObjects that weren't the same size as the squareGridSize were off-centre in the level.
Found and fixed a weird bug that resulted from Bounds.getWidth() returning 'width/2' for Rectangles, but 'width' for everything else. Rectangles can now be used in the level