Laser Ability
As the robot, you can now fire your laser with 'V'
Added keybinds to WASD for movement.
Moved Ability enum into the Player class.
Implemented the currentBoost and currentAbility bits of the UI
Refactored 'SetPrevPosition' to 'SetPrevIndex' to imply that it is in terms of level squares, and not pixels in the window.
The Laser class manages displaying the laser onscreen. Right now there's only support for one laser at a time, and it will return gracefully if there is already a laser onscreen.