Animation Test
Redid most of the LevelObject code so that we don't have to worry about all those dumb off-by-a-bit errors that make things misaligned with the grid.
I used a StackPane to house an invisible square called hitbox, and the model. So now the StackPane is always a fixed size, regardless of what the model does.
Animation Test in progress. Be prepared for nightmare fuel.